[Decomp][SpellMgr] Cohesion split: extract spell data loaders to SpellMgrData.cpp
